code fonctionnel

+++ inc/decryptRS.h
@@ -15,12 +15,12 @@
 #define DECRYPTRS_H
 
 
-void evaluation_fonction(fmpz_t res, fmpz_mod_poly_t fonction, int valeur_x);
+void evaluation_fonction(fmpz_t res, fmpz_mod_poly_t fonction, long valeur_x);
 void xor_polynome(fmpz_mod_poly_t resu, fmpz_mod_poly_t poly1_v, fmpz_mod_poly_t poly2_v);
 void division(fmpz_mod_poly_t q, fmpz_mod_poly_t r, fmpz_mod_poly_t dividente, fmpz_mod_poly_t deviseur);
-void algo_euclide( fmpz_mod_poly_t localisation,fmpz_mod_poly_t amplitude , fmpz_mod_poly_t x2t_v, fmpz_mod_poly_t syndrome_v,int tt);
-bool calcul_poly_syndrome(fmpz_mod_poly_t syndrome, fmpz_mod_poly_t data,int tt);
+void algo_euclide( fmpz_mod_poly_t localisation,fmpz_mod_poly_t amplitude , fmpz_mod_poly_t x2t_v, fmpz_mod_poly_t syndrome_v,long tt);
+bool calcul_poly_syndrome(fmpz_mod_poly_t syndrome, fmpz_mod_poly_t data,long tt);
 void derivation(fmpz_mod_poly_t res, fmpz_mod_poly_t function);
-int decode(fmpz_mod_poly_t data, fmpz_mod_poly_t received ,int tt );
+long decode(fmpz_mod_poly_t data, fmpz_mod_poly_t received ,long tt );
 void test_decode();
 #endif
